Refactor renderer implementations
- Share default implementation in OutputRendering
- Add initial implementation to GitHubRenderer
- Delete unnecessary implementations in TerminalRenderer
