+++ Sources/XcbeautifyLib/Renderers/OutputRendering.swift
@@ -260,27 +260,61 @@ extension OutputRendering {
 }
 
 extension OutputRendering {
-    func formatTestSuiteAllTestsPassed(group: TestSuiteAllTestsPassedCaptureGroup)  -> String? {
+
+    func format(line: String, command: String, pattern: Pattern, arguments: String) -> String? {
+        let template = command.style.Bold + " " + arguments
+
+        guard let formatted = try? NSRegularExpression(pattern: pattern.rawValue)
+                .stringByReplacingMatches(
+                    in: line,
+                    range: NSRange(location: 0, length: line.count),
+                    withTemplate: template)
+            else {
+                return nil
+        }
+
+        return formatted
+    }
+
+    func formatAnalyze(group: AnalyzeCaptureGroup) -> String {
+        let filename = group.fileName
+        let target = group.target
+        return colored ? "[\(target.f.Cyan)] \("Analyzing".s.Bold) \(filename)" : "[\(target)] Analyzing \(filename)"
+    }
+
+    func formatCleanRemove(group: CleanRemoveCaptureGroup) -> String {
+        let directory = group.directory
+        return colored ? "\("Cleaning".s.Bold) \(directory)" : "Cleaning \(directory)"
+    }
+
+    func formatCodeSign(group: CodesignCaptureGroup) -> String {
+        let command = "Signing"
+        let sourceFile = group.file
+        return colored ? command.s.Bold + " " + sourceFile.lastPathComponent : command + " " + sourceFile.lastPathComponent
+    }
+
+    func formatCodeSignFramework(group: CodesignFrameworkCaptureGroup) -> String {
+        let frameworkPath = group.frameworkPath
+        return colored ? "\("Signing".s.Bold) \(frameworkPath)" : "Signing \(frameworkPath)"
+    }
+
+    func formatCompile(group: CompileFileCaptureGroup) -> String {
+        let filename = group.filename
+        let target = group.target
+        return colored ? "[\(target.f.Cyan)] \("Compiling".s.Bold) \(filename)" : "[\(target)] Compiling \(filename)"
+    }
+
+    func formatCompileCommand(group: CompileCommandCaptureGroup) -> String? {
         return nil
     }
 
-    func formatTestSuiteAllTestsFailed(group: TestSuiteAllTestsFailedCaptureGroup) -> String? {
-        return nil
+    func formatCopy(group: CopyCaptureGroup) -> String {
+        let filename = group.file
+        let target = group.target
+        return colored ? "[\(target.f.Cyan)] \("Copying".s.Bold) \(filename)" : "[\(target)] Copying \(filename)"
     }
 
-    func formatTestsRunCompletion(group: TestsRunCompletionCaptureGroup) -> String? {
-        return nil
-    }
-
-    func formatWriteFile(group: WriteFileCaptureGroup) -> String? {
-        return nil
-    }
-
-    func formatWriteAuxiliaryFiles(group: WriteAuxiliaryFilesCaptureGroup) -> String? {
-        return nil
-    }
-
-    func formatShellCommand(group: ShellCommandCaptureGroup) -> String? {
+    func formatCursor(group: CursorCaptureGroup) -> String? {
         return nil
     }
 
@@ -292,16 +326,25 @@ extension OutputRendering {
         return nil
     }
 
-    func formatTestCasesStarted(group: TestCaseStartedCaptureGroup) -> String? {
-        return nil
+    func formatGenerateCoverageData(group: GenerateCoverageDataCaptureGroup) -> String {
+        return colored ? "\("Generating".s.Bold) code coverage data..." : "Generating code coverage data..."
     }
 
-    func formatTIFFUtil(group: TIFFutilCaptureGroup) -> String? {
-        return nil
+    func formatCoverageReport(group: GeneratedCoverageReportCaptureGroup) -> String {
+        let filePath = group.coverageReportFilePath
+        return colored ? "\("Generated".s.Bold) code coverage report: \(filePath.s.Italic)" : "Generated code coverage report: \(filePath)"
     }
 
-    func formatCursor(group: CursorCaptureGroup) -> String? {
-        return nil
+    func formatGenerateDsym(group: GenerateDSYMCaptureGroup) -> String {
+        let dsym = group.dsym
+        let target = group.target
+        return colored ? "[\(target.f.Cyan)] \("Generating".s.Bold) \(dsym)" : "[\(target)] Generating \(dsym)"
+    }
+
+    func formatLibtool(group: LibtoolCaptureGroup) -> String {
+        let filename = group.fileName
+        let target = group.target
+        return colored ? "[\(target.f.Cyan)] \("Building library".s.Bold) \(filename)" : "[\(target)] Building library \(filename)"
     }
 
     func formatLinkerDuplicateSymbolsLocation(group: LinkerDuplicateSymbolsLocationCaptureGroup) -> String? {
@@ -311,4 +354,199 @@ extension OutputRendering {
     func formatLinkerUndefinedSymbolLocation(group: LinkerUndefinedSymbolLocationCaptureGroup) -> String? {
         return nil
     }
+
+    func formatLinking(group: LinkingCaptureGroup) -> String {
+        let target = group.target
+#if os(Linux)
+        return colored ? "[\(target.f.Cyan)] \("Linking".s.Bold)" : "[\(target)] Linking"
+#else
+        let filename = group.binaryFilename
+        return colored ? "[\(target.f.Cyan)] \("Linking".s.Bold) \(filename)" : "[\(target)] Linking \(filename)"
+#endif
+    }
+
+    func formatPackageCheckingOut(group: PackageCheckingOutCaptureGroup) -> String {
+        let version = group.version
+        let package = group.package
+        return colored ? "Checking out " + package.s.Bold + " @ " + version.f.Green : "Checking out \(package) @ \(version)"
+    }
+
+    func formatPackageEnd() -> String {
+        return colored ? "Resolved source packages".s.Bold.f.Green : "Resolved source packages"
+    }
+
+    func formatPackageFetching(group: PackageFetchingCaptureGroup) -> String {
+        let source = group.source
+        return "Fetching " + source
+    }
+
+    func formatPackageItem(group: PackageGraphResolvedItemCaptureGroup) -> String  {
+        let name = group.packageName
+        let url = group.packageURL
+        let version = group.packageVersion
+        return colored ? name.s.Bold.f.Cyan + " - " + url.s.Bold + " @ " + version.f.Green : "\(name) - \(url) @ \(version)"
+    }
+
+    func formatPackageStart() -> String {
+        return colored ? "Resolving Package Graph".s.Bold.f.Cyan : "Resolving Package Graph"
+    }
+
+    func formatPackageUpdating(group: PackageUpdatingCaptureGroup) -> String {
+        let source = group.source
+        return "Updating " + source
+    }
+
+    func formatParallelTestCaseAppKitPassed(group: ParallelTestCaseAppKitPassedCaptureGroup) -> String {
+        let indent = "    "
+        let testCase = group.testCase
+        let time = group.time
+        return colored ? indent + TestStatus.pass.foreground.Green + " " + testCase + " (\(time.coloredTime(colored: colored)) seconds)" : indent + TestStatus.pass + " " + testCase + " (\(time)) seconds)"
+    }
+
+    func formatParallelTestCasePassed(group: ParallelTestCasePassedCaptureGroup) -> String {
+        let indent = "    "
+        let testCase = group.testCase
+        let device = group.device
+        let time = group.time
+        return colored ? indent + TestStatus.pass.foreground.Green + " " + testCase + " on '\(device)' (\(time.coloredTime(colored: colored)) seconds)" : indent + TestStatus.pass + " " + testCase + " on '\(device)' (\(time) seconds)"
+    }
+
+    func formatParallelTestingPassed(line: String, group: ParallelTestingPassedCaptureGroup) -> String {
+        return colored ? line.s.Bold.f.Green : line
+    }
+
+    func formatParallelTestSuiteStarted(group: ParallelTestSuiteStartedCaptureGroup) -> String {
+        let testSuite = group.suite
+        let deviceDescription = " on '\(group.device)'"
+        let heading = "Test Suite \(testSuite) started\(deviceDescription)"
+        return colored ? heading.s.Bold.f.Cyan : heading
+    }
+
+    func formatParallelTestingStarted(line: String, group: ParallelTestingStartedCaptureGroup) -> String {
+        return colored ? line.s.Bold.f.Cyan : line
+    }
+
+    func formatPhaseScriptExecution(group: PhaseScriptExecutionCaptureGroup) -> String {
+        let phaseName = group.phaseName
+        let target = group.target
+        // Strip backslashed added by xcodebuild before spaces in the build phase name
+        let strippedPhaseName = phaseName.replacingOccurrences(of: "\\ ", with: " ")
+        return colored ? "[\(target.f.Cyan)] \("Running script".s.Bold) \(strippedPhaseName)" : "[\(target)] Running script \(strippedPhaseName)"
+    }
+
+    func formatPhaseSuccess(group: PhaseSuccessCaptureGroup) -> String {
+        let phase = group.phase.capitalized
+        return colored ? "\(phase) Succeeded".s.Bold.f.Green : "\(phase) Succeeded"
+    }
+
+    func formatProcessInfoPlist(group: ProcessInfoPlistCaptureGroup) -> String {
+        let plist = group.filename
+
+        if let target = group.target {
+            // Xcode 10+ output
+            return colored ? "[\(target.f.Cyan)] \("Processing".s.Bold) \(plist)" : "[\(target)] \("Processing") \(plist)"
+        } else {
+            // Xcode 9 output
+            return colored ? "Processing".s.Bold + " " + plist : "Processing" + " " + plist
+        }
+    }
+
+    func formatProcessPch(group: ProcessPchCaptureGroup) -> String {
+        let filename = group.file
+        let target = group.buildTarget
+        return colored ? "[\(target.f.Cyan)] \("Processing".s.Bold) \(filename)" : "[\(target)] Processing \(filename)"
+    }
+
+    func formatProcessPchCommand(group: ProcessPchCommandCaptureGroup) -> String {
+        let filePath = group.filePath
+        return colored ? "\("Preprocessing".s.Bold) \(filePath)" : "Preprocessing \(filePath)"
+    }
+
+    func formatShellCommand(group: ShellCommandCaptureGroup) -> String? {
+        return nil
+    }
+
+    func formatTargetCommand(command: String, group: TargetCaptureGroup) -> String {
+        let target = group.target
+        let project = group.project
+        let configuration = group.configuration
+        return colored ? "\(command) target \(target) of project \(project) with configuration \(configuration)".s.Bold.f.Cyan : "\(command) target \(target) of project \(project) with configuration \(configuration)"
+    }
+
+    func formatTestCaseMeasured(group: TestCaseMeasuredCaptureGroup) -> String {
+        let indent = "    "
+        let testCase = group.testCase
+        let name = group.name
+        let unitName = group.unitName
+        let value = group.value
+        let deviation = group.deviation.coloredDeviation(colored: colored)
+
+        let formattedValue: String
+        if unitName == "seconds" {
+            formattedValue = value.coloredTime(colored: colored)
+        } else {
+            formattedValue = value
+        }
+        return indent + (colored ? TestStatus.measure.foreground.Yellow : TestStatus.measure) + " "  + testCase + " measured (\(formattedValue) \(unitName) ±\(deviation)% -- \(name))"
+    }
+
+    func formatTestCasePassed(group: TestCasePassedCaptureGroup) -> String {
+        // TODO: Extract to shared property
+        let indent = "    "
+        let testCase = group.testCase
+        let time = group.time
+        return colored ? indent + TestStatus.pass.foreground.Green + " " + testCase + " (\(time.coloredTime(colored: colored)) seconds)" : indent + TestStatus.pass + " " + testCase + " (\(time) seconds)"
+    }
+
+    func formatTestCasePending(group: TestCasePendingCaptureGroup) -> String {
+        let indent = "    "
+        let testCase = group.testCase
+        return colored ? indent + TestStatus.pending.foreground.Yellow + " "  + testCase + " [PENDING]" : indent + TestStatus.pending + " "  + testCase + " [PENDING]"
+    }
+
+    func formatTestCasesStarted(group: TestCaseStartedCaptureGroup) -> String? {
+        return nil
+    }
+
+    func formatTestsRunCompletion(group: TestsRunCompletionCaptureGroup) -> String? {
+        return nil
+    }
+
+    func formatTestSuiteAllTestsFailed(group: TestSuiteAllTestsFailedCaptureGroup) -> String? {
+        return nil
+    }
+
+    func formatTestSuiteAllTestsPassed(group: TestSuiteAllTestsPassedCaptureGroup)  -> String? {
+        return nil
+    }
+
+    func formatTestSuiteStart(group: TestSuiteStartCaptureGroup) -> String {
+        let testSuite = group.testSuiteName
+        return colored ? testSuite.s.Bold : testSuite
+    }
+
+    func formatTestSuiteStarted(group: TestSuiteStartedCaptureGroup) -> String {
+        let testSuite = group.suite
+        let heading = "Test Suite \(testSuite) started"
+        return colored ? heading.s.Bold.f.Cyan : heading
+    }
+
+    func formatTIFFUtil(group: TIFFutilCaptureGroup) -> String? {
+        return nil
+    }
+
+    func formatTouch(group: TouchCaptureGroup) -> String {
+        let filename = group.filename
+        let target = group.target
+        return colored ? "[\(target.f.Cyan)] \("Touching".s.Bold) \(filename)" : "[\(target)] Touching \(filename)"
+    }
+
+    func formatWriteAuxiliaryFiles(group: WriteAuxiliaryFilesCaptureGroup) -> String? {
+        return nil
+    }
+
+    func formatWriteFile(group: WriteFileCaptureGroup) -> String? {
+        return nil
+    }
+
 }
